+++ packages/@expo/cli/src/utils/env.ts
@@ -136,25 +136,11 @@ class Env {
   get HTTP_PROXY(): string {
     return process.env.HTTP_PROXY || process.env.http_proxy || '';
   }
-<<<<<<< HEAD
-||||||| parent of 5234fe38ff (feature(cli): add custom inspector proxy based on `metro-inspector-proxy` (#21449))
-
-  /** **Experimental:** Prevent Metro from using the `compilerOptions.paths` feature from `tsconfig.json` (or `jsconfig.json`) to enable import aliases. */
-  get EXPO_USE_PATH_ALIASES(): boolean {
-    return boolish('EXPO_USE_PATH_ALIASES', false);
-  }
-=======
-
-  /** **Experimental:** Prevent Metro from using the `compilerOptions.paths` feature from `tsconfig.json` (or `jsconfig.json`) to enable import aliases. */
-  get EXPO_USE_PATH_ALIASES(): boolean {
-    return boolish('EXPO_USE_PATH_ALIASES', false);
-  }
 
   /** **Experimental:** Use the network inspector by overriding the metro inspector proxy with a custom version */
   get EXPO_USE_CUSTOM_INSPECTOR_PROXY(): boolean {
     return boolish('EXPO_USE_CUSTOM_INSPECTOR_PROXY', false);
   }
->>>>>>> 5234fe38ff (feature(cli): add custom inspector proxy based on `metro-inspector-proxy` (#21449))
 }
 
 export const env = new Env();